backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
click==8.1.7
exceptiongroup==1.3.0
fastapi==0.119.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.1.0
mangum==0.19.0
packaging==25.0
pusher==3.3.3
pluggy==1.6.0
pydantic==2.12.0
pydantic_core==2.41.1
Pygments==2.19.2
pytest==8.4.2
sniffio==1.3.1
starlette==0.48.0
tomli==2.3.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
boto3==1.34.77
python-dotenv==1.0.1
stream-chat==4.26.0
openai>=1.42.0
